refactor: move controllers and usecases classes to folders specific to confirmation email
